Move verify/unverify instr builders to their own mod
